Retry db transactions the right way
Previously we were retrying db transactions by reusing the same
transaction. This doesn't work when CRDB failed the transaction, so
ensure we create a new transaction each time we're trying things.

This also allows the maximum number of retries and the maximum interval
between retries to be configured via cmdline/env vars.