Split delete operation into two separate transactions
We were checking whether the current delete operation was deleting the
last datum of metadata or userdata, and only if that was the case would
we delete instance ip addresses from the db. But in practice the delete
operations for metadata and userdata would come in around the same time,
resulting in a race condition if the delete operation of one thread took
longer than the other to complete.

This separates the decision to delete instance ip addresses to *after*
the current delete operation has finished, ensuring that the last delete
thread to run will see both datums are gone and perform the ip address
deletion.