Refactor upserts for instance ip addresses
Enable the use of SELECT FOR UPDATE when finding existing instance ip
address rows that may be changed during an upsert transaction. This will
lock the rows, even from reads, until the transaction is commit. Based
on the combination of this and the previous commit where we enforce ip
address uniqueness in the db, we should have good protection from upsert
race conditions.

Also refactor the retry logic into its own wrapper function for
improved clarity.