Add authenticated refresh endpoints for metadata and userdata
These endpoints are intended for admin/operators to be able to force a
refresh of metadata or userdata for troubleshooting purposes. Regardless
of whether the data is in the local database, these endpoints invoke the
lookup client to pull and save the latest data from the upstream source.

Also includes misc cleanups to make the linter happy.

Signed-off-by: Scott Garman <sgarman@equinix.com>