Refactoring of upserts to fix race conditions
When two metadata updates arrive in close succession, it is possible for
the first one to overwrite the second one if the first transaction is
retried.

This refactors our doUpsert code to be more robust against race
conditions in two ways:

* We use SELECT FOR UPDATE to lock all ip address rows at the start of
  the transaction
* We use crdb.ExecuteTx(), which automatically handles retries due to
  transient errors (e.g, db serialization errors).

The outer retry logic (for loop) is only there as a catch-all for edge
cases that need to be better understood, because it too still allows for
race conditions if the transaction run by ExecuteTx() is failing for
transient reasons. We log this case thoroughly to allow for debugging.

Signed-off-by: Scott Garman <sgarman@equinix.com>