Prevent stale metadata upserts from being applied
If metadata contains the optional field `updated_at`, interpret this
field as a timestamp representing when the metadata was generated. With
this info, we can prevent some race conditions when out of order updates
are received. Since incoming metadata upserts include all relevant data
(IPs and current metadata in full), stale updates can simply be
discarded.

Signed-off-by: Scott Garman <sgarman@equinix.com>