Lower db max retries from 5 to 3
At this point, the only times we need to restart db transactions from
scratch are when they get canceled due to a context timeout. If this
doesn't get sorted out after 3 retries, 5 is just going to prolong the
db locks and cascade to other problems. So pare this back a bit.

Signed-off-by: Scott Garman <sgarman@equinix.com>